requirements-dev.txt
-r requirements.txt
black~=19.10b0
isort~=4.3.21
mypy~=0.770
mypy-extensions~=0.4.3
pyre-check~=0.0.46
pyre-extensions~=0.0.18
typing-extensions~=3.7.4.2
typing-inspect~=0.6.0
sqlalchemy-stubs~=0.3
